requirements.txt
mysql-connector-python
python-dotenv
langchain-community
langchain_google_genai
pypdf